Pull single/array codegen type factories out into their own files
